Task 11: 101-print_number.c

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -1,25 +1,26 @@
 #include "main.h"
 
 /**
- * print_number - Function that prints an integer.
- * @n: int type number
- * Description: Can only use _putchar to print.
+ * print_number - Prints an integer.
+ * @n: The integer to be printed
+ *
+ * Description: This function can only use _putchar to print.
  */
 void print_number(int n)
 {
-	long m; /* power of 10 */
+	int m; /* power of 10 */
 	int c; /* boolean check */
-	long num; /* convert int to long */
+	int num; /* convert int to long */
 
 	num = n;
-	/* negatives */
+	/* Handle negatives */
 	if (num < 0)
 	{
 		num *= -1;
 		_putchar('-');
 	}
 
-	/* count up */
+	/* Count up */
 	m = 1;
 	c = 1;
 	while (c)
@@ -30,7 +31,7 @@ void print_number(int n)
 			c = 0;
 	}
 
-	/* count down */
+	/* Count down */
 	while (num >= 0)
 	{
 		if (m == 1)
